pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot made using discord.py"
authors = ["Abdulmalik Tarbooly"]
license = "MIT"
readme = "README.md"
packages = [{include = "discord_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.2.2"
python-dotenv = "^1.0.0"
py-cord = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
